chore: delete the collab server, its tables, and every wire to it
Nobody ever used what it uniquely provided: zero pages co-edited by two
humans in product history, zero pastes ever live-edited. What it did
provide: a second source of truth that fought agent writes, a deploy
bounce that dropped every open editor several times a day, and the two
editor data-corruption bugs of the 2026-08-06 incident. The editor now
loads and saves directly (#981), so:

- collab/ is gone, along with its Render/compose service, Caddy route,
  start.sh block, GHCR image, CI audit steps, and frontend deps
  (@hocuspocus/provider, yjs, tiptap collaboration extensions).
- /api/v1/collab/* authorize endpoints removed; paste live-edit auth
  removed (token-authed REST paste editing is unaffected).
- collab_projection removed from the page PATCH: the guard-bypass existed
  only for the editor's own projection saves, and every writer now goes
  through optimistic concurrency.
- Migration 0181 drops page_collab_documents + paste_collab_documents.
  content_markdown always carried every edit's export, so nothing is lost.
- Onboarding/email copy no longer promises real-time co-editing.

Deploy order: merge only after #981 is live (the editor must stop calling
the server before the server dies). The Render service stash_collab is
dashboard-managed — suspend/delete it manually after this deploys.
