+++ Caddyfile
@@ -13,11 +13,6 @@ app.example.com {
         reverse_proxy backend:3456
     }
 
-    # Yjs collaboration WebSocket
-    handle /collab* {
-        reverse_proxy collab:3458
-    }
-
     # Frontend
     handle {
         reverse_proxy frontend:3457
