--- collab/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ES2022",
-    "module": "NodeNext",
-    "moduleResolution": "NodeNext",
-    "strict": true,
-    "esModuleInterop": true,
-    "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true,
-    "outDir": "dist",
-    "lib": ["ES2022", "DOM"]
-  },
-  "include": ["src/**/*.ts"]
-}
