+++ docker-compose.prod.yml
@@ -14,7 +14,7 @@
 #   docker compose -f docker-compose.prod.yml -f docker-compose.local.yml pull
 #   docker compose -f docker-compose.prod.yml -f docker-compose.local.yml up -d
 #
-# Backend, frontend, and collab use pre-built images from GHCR.
+# Backend and frontend use pre-built images from GHCR.
 #
 # Required env vars in .env:
 #   POSTGRES_USER, POSTGRES_PASSWORD, POSTGRES_DB
@@ -126,21 +126,6 @@ services:
     environment:
       BACKEND_INTERNAL_URL: http://backend:3456
 
-  collab:
-    image: ghcr.io/fergana-labs/stash-collab:0.1.362
-    restart: always
-    expose:
-      - "3458"
-    depends_on:
-      postgres:
-        condition: service_healthy
-      backend:
-        condition: service_started
-    env_file: .env
-    environment:
-      DATABASE_URL: postgresql://${POSTGRES_USER}:${POSTGRES_PASSWORD}@postgres:5432/${POSTGRES_DB}
-      BACKEND_URL: http://backend:3456
-
   caddy:
     image: caddy:2-alpine
     restart: always
@@ -154,7 +139,6 @@ services:
     depends_on:
       - backend
       - frontend
-      - collab
 
 volumes:
   postgres_data:
