+++ frontend/package.json
@@ -12,9 +12,6 @@
   },
   "dependencies": {
     "@auth0/nextjs-auth0": "^4.22.0",
-    "@hocuspocus/provider": "^4.0.0",
-    "@tiptap/extension-collaboration": "^3.6.6",
-    "@tiptap/extension-collaboration-caret": "^3.6.6",
     "@tiptap/extension-image": "^3.22.2",
     "@tiptap/extension-link": "3.6.6",
     "@tiptap/extension-placeholder": "3.6.6",
@@ -57,7 +54,6 @@
     "tippy.js": "^6.3.7",
     "tw-animate-css": "^1.4.0",
     "y-protocols": "^1.0.6",
-    "yjs": "^13.6.23",
     "zustand": "^5.0.14"
   },
   "devDependencies": {
