+++ frontend/src/app/globals.css
@@ -477,36 +477,6 @@ details > summary::-webkit-details-marker {
   border-right: none;
 }
 
-/* Collaboration cursor styles */
-.collaboration-cursor__caret {
-  border-left: 1px solid #0d0d0d;
-  border-right: 1px solid #0d0d0d;
-  margin-left: -1px;
-  margin-right: -1px;
-  pointer-events: none;
-  position: relative;
-  word-break: normal;
-}
-
-.collaboration-cursor__label {
-  border-radius: 3px 3px 3px 0;
-  color: #fff;
-  font-size: 11px;
-  font-weight: 600;
-  left: -1px;
-  line-height: normal;
-  padding: 0.1rem 0.3rem;
-  position: absolute;
-  top: -1.4em;
-  user-select: none;
-  white-space: nowrap;
-  font-family: var(--font-mono), monospace;
-}
-
-.collaboration-selection {
-  border-radius: 2px;
-}
-
 /* Task list checkbox styles */
 .tiptap ul[data-type="taskList"] {
   list-style: none;
